feat: 메뉴 리스트 마크업
- 더미 데이터파일 생성
- 기본 마크업
